Partial fix for https://bashoeng.atlassian.net/browse/RTS-1094
In this world an attempt is made to insert values into a non-existant table.

This itself is a bug - the existance of the table should be checked - instead a crash happens

This crash message is returned to the connection server in riak_shell which then crashes.

This fix makes riak-shell well behaved in the presence of unknown crashes server side
with a push to the user to report the bug to Basho.

The actual bug that triggered it will be fixed in a different PR