Bug fix for unicode in riak-shell
Unicode in input was being handled correctly by the riak_shell
lexer/parser (and by the SQL one) but not by the cmd pretty printer

This is used in writing the history and the logs and it was borking
on unicode - in particular smart quotes inserted into SQL on pasting
from some editors.

The regular expression clean up and output munging is now unicode
friendly.

riak_shell doesn't log errored commands so it is hard to regression
test for this.